+++ src/util.h
@@ -11,7 +11,7 @@ int _GetFileTime(const char *fileName, char *lpszString, long dwSize);
 #endif
 
 uint32_t hashlittle(const void *key, size_t length, uint32_t initval);
-uint32_t HashFile(const char *file, int *mtime, size_t *size);
+uint32_t HashFile(const char *file, int *mtime, uint64_t *size);
 int TouchFile(const char *fileName);
 void hex_dump(void *data, int size);
